agentic.ai/requirements.txt
# FastAPI ve temel bağımlılıklar
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain ve LLM entegrasyonu
langchain==0.1.0
langchain-google-genai==0.0.5

# Google Gemini
google-generativeai>=0.3.1,<0.4.0

# HTTP istekleri için
requests==2.31.0
aiohttp==3.9.1

# Async desteği
asyncio
aiofiles==23.2.1

# Logging ve yardımcı araçlar
python-multipart==0.0.6
python-dotenv==1.0.0

# Test için
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Sistem metrikleri için
psutil==5.9.6 